requirements.txt
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
idna==3.10
kagglehub==0.3.13
numpy==2.3.3
packaging==25.0
pandas==2.3.3
protobuf==6.32.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
text-unidecode==1.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
